feat(forum): renaming www app
